# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Crudtodo",
    "description": "Requre authentication for all routes, except for /login, /register and /logout. Only user can delete their own account and their contents.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/LETHANHDAT1993/CrudTODO/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "http://localhost:3000",
      "description": "SwaggerHub API CrudTodo - SQL Database"
    },
    {
      "url": "http://localhost:3000",
      "description": "SwaggerHubSwaggerHub API CrudTodo - SQL Database"
    }
  ],
  "paths": {
...
    "/sql/user/{id}": {
      "get": {
        "tags": [
          "Require authentication"
        ],
        "description": "Read one user's infor",
        "operationId": "readOneUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request are accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/oneUserRes"
                }
              }
            }
          },
          "401": {
            "description": "Block by middleware",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageRes"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete user"
        ],
        "description": "If the request method is delete, the ID must be the same to req.session.userId.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request are accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageRes"
                }
              }
            }
          },
          "401": {
            "description": "Block by middleware",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/messageRes"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "oneUserRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "description": "Created time",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "description": "Udated time",
            "format": "date-time"
          }
        }
      },
...
      "messageRes": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}